Update patch set 1
Patch Set 1:

I tried so hard to make it work that way, but the maps_is_readable call is only compiled once, so I can't get one version for the remote calls, and one for the local.

I think there is a possible way to make it work better, but I want to try and get upstream to have a better solution so I can get rid of all of this.

Patch-set: 1