Update patch set 8
Patch Set 8:

>I've got to ask, how is this a workaround? It just makes getcontext always return an error which means it's not working.

Yes, initially it was a build workaround. Wouldn't you prefer to have the calling code fail if the feature is unimplemented? Unless an empty implementation is a correct one...

>Are you going to create a getcontext like the one for arm, or is there some other way that you plan to implement this in aarch64?

No plans for now. I haven't yet seen a place where this had proven to be necessary.

>By the way, I submitted a patch to upstream libunwind to change the %x to %lx. Once it's accepted (I hope), I'll pull it back in.

Ok.

Patch-set: 8
Label: Verified=0