Update patch set 1
Patch Set 1:

i wonder if instead there should be some check when subtracting 4 that that doesn't take us out of the function we're currently in?

with this patch we successfully unwind, but dladdr on the result of _Unwind_GetIP says it doesn't know what function in the vdso we're in, whereas libgcc correctly says __kernel_rt_sigreturn (though all its other returned PC values point to the next instruction; you really can't win!).

Patch-set: 1