Update patch set 1
Patch Set 1:

Ok, so they do use offsetof :-), but in a separate program.

I got the 0xA8 manually. gen-offsets spits out 0xB0 because the uc_mcontext field gets aligned to 16 bytes. This is what caught me.

Apparently gcc and clang align structures in structures to the maximum field alignment of the inner structure (determined by experimentation). Here there is a __reserved field in struct sigcontext which is aligned to 16 bytes.

According to the standard this is perfectly valid to do (1421 Each non-bit-field member of a structure or union object is aligned in an implementation-defined manner appropriate to its type).

Any manual calculation is really hopeless...

Since Elliot fixed the padding in bionic a few month ago I think the conclusion is that no change is needed.

Patch-set: 1