Update patch set 1
Patch Set 1:

Are you planning to use this for unwinding?

If so, please don't. If you can use system/core/libbacktrace, or describe why you can't use it, that would really be useful.

There is no guarantee that this will continue to exist in the future, while libbacktrace will continue.

Patch-set: 1