Update patch set 3
Patch Set 3:

Small rework. I tested this and it passes one of the stack unwinding tests, but not the other.

stack_unwinding.unwind_through_signal_frame fails
stack_unwinding.unwind_through_signal_frame_SA_SIGINFO passes

On the emulator, when using the alternate stack, the signal context data is not complete so it's impossible to unwind anything. It also means that thread unwinding on mips is broken.

Is this something that is due to the kernel used in the mips emulator, or something else?

Patch-set: 3