Implement method to read elf data from memory.
I rewrote the entirety of elfxx.h/elfxx.c to support this.

For the rewrite, I fixed the way

In addition, I only had to duplicate the functionality from one
function for doing dwarf unwinding. I also added checks for
the symbol table read that verifies we're not going to read past the
end of the elf image.

Bug: 19517541

Change-Id: I6075d87035dca41fa5e5dfcbe287a680a7671836