Decompress debug symbols.
Debug symbols may be stored in LZMA compressed section .gnu_debugdata.

libunwind already contained some support for it however it was
disabled at compile time.  This CL re-enables it and extends it.

The decompression code is based on LZMA SDK.

It also re-enables support for .debug_frame on all platforms
(it was previously only enabled on ARM).  The .debug_frame
can be loaded from the compressed section as well now.

Change-Id: Ie199ce1fc31770940db288b5e57b50942f12be7b