Update patch set 1
Patch Set 1:

With those I pass:
  mm -j40 test-art-host-gtest
  m -j32 test-art-host-run-test
  test/run-test --host --gdb 137
  test/run-test --host --gdb --64 137

What about the targets? At least arm32 and arm64 are important (so testing on N9 is easiest). I know we'd expect it to work, but better safe than sorry... :-)

 > How do I run the debuggerd/libbacktrace tests?

Chris?

The libbacktrace test is a gtest. Run mm in system/core/libbacktrace, and it should produce a binary, I believe.

I'm not sure where the debuggerd tests are (or if there's only libbacktrace testing).

Patch-set: 1