Update patch set 1
Patch Set 1:

If crasher hangs, then that means debuggerd is hanging.

However, I see in aosp that bluetooth is constantly crashing which means that if another process is trying to crash, it might get a bit stuck. So if you see a constant stream of crashes, do a stop and then try crasher again.

Patch-set: 1