Update patch set 4
Patch Set 4:

This actually does not compile on N9.
After several hours of effort, I still do not seem to be able to convince the linker to properly link to the LZMA library.
Just enabling libunwind_debug without this CL does not seem to work either.
I think the trouble is with the setup of the second (64-bit) target architecture, but I am not sure what the exact problem is.  A bit of help would be greatly appreciated.

Patch-set: 4