Update patch set 2
Patch Set 2:

Specifically, this was only needed if someone linked the static libunwind along with the dynamic version. Dan can you confirm that this isn't done any more, and that there is no plan to do this kind of thing in the future.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0