Update patch set 1
Patch Set 1: Code-Review-1

I don't really like this. You are exporting internal functions that might or might not rely on internal data structures.

I also don't like the way simpleperf code is directly accessing libunwind like this.

When we replace this, we'll have to rewrite all of this code.

I would much prefer that something be added to libbacktrace, and then called through that. Then when having to rewrite, the only change is in libbacktrace.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1