Update patch set 2
Patch Set 2: -Code-Review

Everything looks good, this does allow unwinding through dlopen.

This is only part one of probably a three part change.

Next up is to add the soname to the apk so it's easier to figure out what is going on.

After that, fix the elf cache to work on these dlopen issues.

Patch-set: 2
Label: Code-Review=0