Force memory unwinds to use a zero map offset.
The dlopen of a shared library from an apk results in huge non-zero
map offsets. This can cause failures due to the checks in the code
of the map offset against the elf data.

Currently, unwinds only work through a map that is in the process's
memory, so this will cover any attempt to unwind through a shared
library that was loaded from an apk.

Bug: 23348999
Change-Id: I3f9d95262be1b17a8adb7e1e4ef0196dabe77d51