Update patch set 2
Patch Set 2:

This happens when trying to unwind through something in an oat file. This winds up with ret having an error value, but dwarf.ip is non-zero probably because it did a partial unwind. So the return code gets converted to a 1 and that bogus dwarf.ip value is thought to be a valid frame.

The next time unw_step is called, usually, we wind up trying to access some garbage and it causes the unwind to stop.

So yes, this is the cause of the bogus extra frame we've been seeing on x86. When I ran with this, I looked and I couldn't see any of weird extra frames any more.

Patch-set: 2