Update patch set 2
Patch Set 2: Code-Review-1

Under what conditions is the memory callback null? If it is, there are many, many places in this code that you would crash.

If this is a real problem, then some other cause needs to be fixed.

Patch-set: 2
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1