Update patch set 1
Patch Set 1:

The pc I was storing before was actually the pc of the calling function, not the current function. If the calling function unwind information and the current unwind information is the same, then there are no problems. However, if the current function modifies the sp in any way, then there is no way that the unwind information will match because the sp does not match what is expected, and any references to values relative to the cfa will get the wrong data (which is what is happening in the failure cases).

Patch-set: 1