Update patch set 1
Patch Set 1:

I see. Is below understanding correct?
pc value is important because:
The first dwarf.ip is pc value.
And pc value is used to determine we should execute dwarf cfa instructions to which position.

Before this change, we are using x30 as pc value, so:
the first dwarf.ip is actually in the function calling current function. (Because unw_tdep_getcontext() is a macro, current function is the one calling unw_tdep_getcontext())

And we are actually executing dwarf cfa instructions for the wrong function.

Patch-set: 1
Reviewer: Gerrit User 1056364 <1056364@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0