[libunwind]: Only add the maps with specials flags to the maps list.
Only add the maps with PROT_EXEC、PROT_EXEC and
non-MAP_FLAGS_DEVICE_MEM flags to the maps list.
This avoids problems trying to open the file associated
with the map that can cause a deadlock if the file is
associated with a device and the driver is single-threaded.

This can also improve the speed of creating maps list,
reducing memory usage.

Change-Id: I966859366ba608e5df726386555cd0f469081a00
Signed-off-by: songjinshi <songjinshi@xiaomi.com>