libunwind is not directly available to vendors
This lib is behind libbacktrace which is again behind libutils.so. This
lib is a member of VNDK but is not intented to be used directly by
vendors.

Furthermore, libunwindbacktrace no longer includes libunwind. This is
because libunwindbacktrace which is a vendor_available:true library
cannot depend on libunwind which is not available to vendors. Instead of
including libunwind inside libunwindbacktrace, libunwind is explicitly
added to the dependency list along with libunwindbacktrace when
libunwindbacktrace is required by other modules. (see libcompiler_rt)

Bug: 64730695
Test: build 2017 pixel device
Change-Id: I3fd2eb25efa26dc5fc55d7e64c8f4e9ab13156ea